test(BpmnAwareTests): #102 split tests to have a single test class per helper method
