+++ camunda-bpm-assert/src/test/java/org/camunda/bpm/engine/test/assertions/bpmn/BpmnAwareTestsCompleteExternalTaskTest.java
@@ -1,5 +1,6 @@
 package org.camunda.bpm.engine.test.assertions.bpmn;
 
+import org.assertj.core.api.Assertions;
 import org.camunda.bpm.engine.externaltask.LockedExternalTask;
 import org.camunda.bpm.engine.runtime.ProcessInstance;
 import org.camunda.bpm.engine.test.Deployment;
@@ -9,10 +10,13 @@ import org.junit.Test;
 
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.fail;
 import static org.camunda.bpm.engine.test.assertions.bpmn.BpmnAwareAssertions.assertThat;
-import static org.camunda.bpm.engine.test.assertions.bpmn.BpmnAwareTests.*;
+import static org.camunda.bpm.engine.test.assertions.bpmn.BpmnAwareTests.completeExternalTask;
+import static org.camunda.bpm.engine.test.assertions.bpmn.BpmnAwareTests.fetchAndLock;
+import static org.camunda.bpm.engine.test.assertions.bpmn.BpmnAwareTests.runtimeService;
 
-public class ExternalTaskHelperTest {
+public class BpmnAwareTestsCompleteExternalTaskTest {
 
   public static final String WORKER_ID = "testWorker";
   public static final String TOPIC = "testTopic";
@@ -21,40 +25,11 @@ public class ExternalTaskHelperTest {
   public ProcessEngineRule processEngineRule = new ProcessEngineRule();
 
   @Test
-  @Deployment(resources = {"bpmn/ExternalTaskHelperTest-fetchAndLock.bpmn"})
-  public void shouldFetchAndLockASingleExternalTaskByTopic() {
-    // Given
-    final ProcessInstance processInstance = runtimeService().startProcessInstanceByKey(
-      "ExternalTaskHelperTest-fetchAndLock"
-    );
-
-    // When
-    List<LockedExternalTask> lockedExternalTasks = fetchAndLock(WORKER_ID, TOPIC, 1);
-
-    // Then
-    assertThat(lockedExternalTasks).hasSize(1);
-    assertThat(lockedExternalTasks.get(0).getWorkerId()).isEqualTo(WORKER_ID);
-  }
-
-  @Test
-  public void fetchAndLockShouldThrowIllegalArgumentException() {
-    try {
-      fetchAndLock(null, TOPIC, 1);
-      fail("IllegalArgumentException expected as workerId is null");
-    } catch (IllegalArgumentException e) {
-      assertThat(e).hasMessageContaining("Illegal call of fetchAndLock(workerId = 'null', topic = 'testTopic') - must" +
-        " not be" +
-        " " +
-        "null");
-    }
-  }
-
-  @Test
-  @Deployment(resources = {"bpmn/ExternalTaskHelperTest-fetchAndLock.bpmn"})
+  @Deployment(resources = {"bpmn/BpmnAwareTestsFetchAndLockTest-completeExternalTask.bpmn"})
   public void shouldCompleteASingleExternalTask() {
     // Given
     final ProcessInstance processInstance = runtimeService().startProcessInstanceByKey(
-      "ExternalTaskHelperTest-fetchAndLock"
+      "BpmnAwareTestsFetchAndLockTest-completeExternalTask"
     );
     List<LockedExternalTask> lockedExternalTasks = fetchAndLock(WORKER_ID, TOPIC, 1);
     assertThat(processInstance).hasNotPassed("externalTask");
@@ -73,8 +48,7 @@ public class ExternalTaskHelperTest {
       completeExternalTask(lockedExternalTask);
       fail("IllegalArgumentException expected as workerId is null");
     } catch (IllegalArgumentException e) {
-      assertThat(e).hasMessageContaining("Illegal call of complete(lockedExternalTask = 'null') - must not be null");
+      Assertions.assertThat(e).hasMessageContaining("Illegal call of complete(lockedExternalTask = 'null') - must not be null");
     }
   }
-
 }
