+++ camunda-bpm-assert/src/test/resources/bpmn/BpmnAwareTestsFetchAndLockTest-completeExternalTask.bpmn
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" id="Definitions_1" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="1.8.2">
+  <bpmn:process id="BpmnAwareTestsFetchAndLockTest-completeExternalTask" name="ExternalTaskHelperTest-fetchAndLock" isExecutable="true">
+    <bpmn:startEvent id="StartEvent_1">
+      <bpmn:outgoing>SequenceFlow_1mh3ry0</bpmn:outgoing>
+    </bpmn:startEvent>
+    <bpmn:endEvent id="EndEvent_0ius2md">
+      <bpmn:incoming>SequenceFlow_1sphqrc</bpmn:incoming>
+    </bpmn:endEvent>
+    <bpmn:sequenceFlow id="SequenceFlow_1sphqrc" sourceRef="externalTask" targetRef="EndEvent_0ius2md" />
+    <bpmn:sequenceFlow id="SequenceFlow_1mh3ry0" sourceRef="StartEvent_1" targetRef="externalTask" />
+    <bpmn:serviceTask id="externalTask" name="Do some work" camunda:type="external" camunda:topic="testTopic">
+      <bpmn:incoming>SequenceFlow_1mh3ry0</bpmn:incoming>
+      <bpmn:outgoing>SequenceFlow_1sphqrc</bpmn:outgoing>
+    </bpmn:serviceTask>
+  </bpmn:process>
+  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="BpmnAwareTestsFetchAndLockTest-completeExternalTask">
+      <bpmndi:BPMNShape id="_BPMNShape_StartEvent_2" bpmnElement="StartEvent_1">
+        <dc:Bounds x="173" y="102" width="36" height="36" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="EndEvent_0ius2md_di" bpmnElement="EndEvent_0ius2md">
+        <dc:Bounds x="417" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="435" y="141" width="0" height="13" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="SequenceFlow_1sphqrc_di" bpmnElement="SequenceFlow_1sphqrc">
+        <di:waypoint xsi:type="dc:Point" x="360" y="120" />
+        <di:waypoint xsi:type="dc:Point" x="417" y="120" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="388.5" y="98.5" width="0" height="13" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_1mh3ry0_di" bpmnElement="SequenceFlow_1mh3ry0">
+        <di:waypoint xsi:type="dc:Point" x="209" y="120" />
+        <di:waypoint xsi:type="dc:Point" x="260" y="120" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="234.5" y="98.5" width="0" height="13" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNShape id="ServiceTask_132yg6x_di" bpmnElement="externalTask">
+        <dc:Bounds x="260" y="80" width="100" height="80" />
+      </bpmndi:BPMNShape>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</bpmn:definitions>
