+++ camunda-bpm-assert/src/test/resources/bpmn/BpmnAwareTestsFetchAndLockTest-fetchAndLock.bpmn
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" id="Definitions_1" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="1.8.2">
-  <bpmn:process id="ExternalTaskHelperTest-fetchAndLock" name="ExternalTaskHelperTest-fetchAndLock" isExecutable="true">
+  <bpmn:process id="BpmnAwareTestsFetchAndLockTest-fetchAndLock" name="ExternalTaskHelperTest-fetchAndLock" isExecutable="true">
     <bpmn:startEvent id="StartEvent_1">
       <bpmn:outgoing>SequenceFlow_1mh3ry0</bpmn:outgoing>
     </bpmn:startEvent>
@@ -15,7 +15,7 @@
     </bpmn:serviceTask>
   </bpmn:process>
   <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="ExternalTaskHelperTest-fetchAndLock">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="BpmnAwareTestsFetchAndLockTest-fetchAndLock">
       <bpmndi:BPMNShape id="_BPMNShape_StartEvent_2" bpmnElement="StartEvent_1">
         <dc:Bounds x="173" y="102" width="36" height="36" />
       </bpmndi:BPMNShape>
